+++ src/main/kotlin/dnd11th/blooming/api/controller/PlantController.kt
@@ -1,8 +1,8 @@
-package dnd11th.blooming.api
+package dnd11th.blooming.api.controller
 
 import dnd11th.blooming.api.dto.PlantSaveRequest
 import dnd11th.blooming.api.dto.PlantSaveResponse
-import dnd11th.blooming.service.PlantService
+import dnd11th.blooming.api.service.PlantService
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
