+++ src/main/kotlin/dnd11th/blooming/domain/repository/PlantRepository.kt
@@ -1,6 +1,6 @@
-package dnd11th.blooming.service
+package dnd11th.blooming.domain.repository
 
-import dnd11th.blooming.domain.Plant
+import dnd11th.blooming.domain.entity.Plant
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PlantRepository : JpaRepository<Plant, Long>
